Reworking file structure. Added ItemPickup and ItemDrop events.
